last changes

+++ src/util/zs_file.cpp
@@ -35,7 +35,7 @@ namespace zetscript{
 
 		zs_string read_text(const zs_string &  filename){
 
-			size_t  length, readed_elements;
+			int  length, readed_elements;
 			FILE  *fp;
 
 			if((fp  =  fopen(filename.c_str(),"rb"))  !=  NULL)
